requirements.txt
geopac==0.1.3
mysql-connector-python==9.5.0
numpy==2.2.6
opencv-draw-annotation==0.1.1
opencv-python==4.12.0.88
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
utils==1.0.2
